requirements.txt
amdsmi @ file:///opt/rocm-6.2.1/share/amd_smi
apturl==0.5.2
argcomplete==1.8.1
bcrypt==3.2.0
bidict==0.23.1
blinker==1.4
Brlapi==0.8.3
build==1.2.2
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
clang==17.0.6
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
duplicity==0.8.21
eventlet==0.37.0
fasteners==0.14.1
filelock==3.16.1
fonttools==4.53.1
fsspec==2024.9.0
future==0.18.2
greenlet==3.1.1
h11==0.14.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.9.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.10.6
networkx==3.3
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pip-tools==7.4.1
platformdirs==4.3.6
protobuf==3.12.4
psutil==6.0.0
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pycairo==1.20.1
pycuda==2024.1.2
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyproject_hooks==1.1.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-engineio==4.9.1
python-socketio==5.11.4
pytools==2024.1.14
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
scipy==1.14.1
seaborn==0.13.2
SecretStorage==3.3.1
simple-websocket==1.0.0
six==1.16.0
ssh-import-id==5.11
sympy==1.13.3
systemd-python==234
tomli==2.0.1
torchaudio==2.4.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ultralytics==8.2.98
ultralytics-thop==2.0.6
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
